more work on win16 build
Needs to include winsock.h/defs.h even for non-win16 funcs!
This is because if they call any of win16 funcs, there should
be a proper declspec, or the mangling will be different and
we'll get undefined symbols.

Especially important for modules that export some non-win16 func
that is called by some win16 func.